+++ tests/execution_engine/split_and_sample/test_sparkdf_execution_engine_splitting.py
@@ -1,9 +1,20 @@
+import datetime
+import os
 from typing import List, Tuple
 from unittest import mock
 
+import numpy as np
+import pandas as pd
 import pytest
 
-from great_expectations.core.batch_spec import RuntimeDataBatchSpec
+import great_expectations.exceptions as ge_exceptions
+from great_expectations.core.batch_spec import (
+    AzureBatchSpec,
+    GCSBatchSpec,
+    PathBatchSpec,
+    RuntimeDataBatchSpec,
+    S3BatchSpec,
+)
 from great_expectations.execution_engine.split_and_sample.data_splitter import DatePart
 from great_expectations.execution_engine.split_and_sample.sparkdf_data_splitter import (
     SparkDataSplitter,
@@ -263,3 +274,319 @@ def test_get_splitter_method(underscore_prefix: str, splitter_method_name: str):
     assert data_splitter.get_splitter_method(
         splitter_method_name_with_prefix
     ) == getattr(data_splitter, splitter_method_name)
+
+
+def test_get_batch_with_split_on_whole_table_runtime(
+    test_sparkdf, basic_spark_df_execution_engine
+):
+    test_sparkdf = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf, splitter_method="_split_on_whole_table"
+        )
+    ).dataframe
+    assert test_sparkdf.count() == 120
+    assert len(test_sparkdf.columns) == 10
+
+
+def test_get_batch_with_split_on_whole_table_filesystem(
+    test_folder_connection_path_csv, basic_spark_df_execution_engine
+):
+    # reader_method not configured because spark will configure own reader by default
+    test_sparkdf = basic_spark_df_execution_engine.get_batch_data(
+        PathBatchSpec(
+            path=os.path.join(test_folder_connection_path_csv, "test.csv"),
+            splitter_method="_split_on_whole_table",
+        )
+    ).dataframe
+    assert test_sparkdf.count() == 6
+    assert len(test_sparkdf.columns) == 2
+
+
+def test_get_batch_with_split_on_whole_table_s3(
+    spark_session, basic_spark_df_execution_engine
+):
+    # noinspection PyUnusedLocal
+    def mocked_get_reader_function(*args, **kwargs):
+        # noinspection PyUnusedLocal,PyShadowingNames
+        def mocked_reader_function(*args, **kwargs):
+            pd_df = pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]})
+            df = spark_session.createDataFrame(
+                [
+                    tuple(
+                        None if isinstance(x, (float, int)) and np.isnan(x) else x
+                        for x in record.tolist()
+                    )
+                    for record in pd_df.to_records(index=False)
+                ],
+                pd_df.columns.tolist(),
+            )
+            return df
+
+        return mocked_reader_function
+
+    spark_engine = basic_spark_df_execution_engine
+    spark_engine._get_reader_fn = mocked_get_reader_function
+
+    test_sparkdf = spark_engine.get_batch_data(
+        S3BatchSpec(
+            path="s3://bucket/test/test.csv",
+            reader_method="csv",
+            reader_options={"header": True},
+            splitter_method="_split_on_whole_table",
+        )
+    ).dataframe
+    assert test_sparkdf.count() == 4
+    assert len(test_sparkdf.columns) == 2
+
+
+def test_get_batch_with_split_on_whole_table_azure(
+    spark_session, basic_spark_df_execution_engine
+):
+    # noinspection PyUnusedLocal
+    def mocked_get_reader_function(*args, **kwargs):
+        # noinspection PyUnusedLocal,PyShadowingNames
+        def mocked_reader_function(*args, **kwargs):
+            pd_df = pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]})
+            df = spark_session.createDataFrame(
+                [
+                    tuple(
+                        None if isinstance(x, (float, int)) and np.isnan(x) else x
+                        for x in record.tolist()
+                    )
+                    for record in pd_df.to_records(index=False)
+                ],
+                pd_df.columns.tolist(),
+            )
+            return df
+
+        return mocked_reader_function
+
+    spark_engine = basic_spark_df_execution_engine
+    spark_engine._get_reader_fn = mocked_get_reader_function
+
+    test_sparkdf = spark_engine.get_batch_data(
+        AzureBatchSpec(
+            path="wasbs://test_container@test_account.blob.core.windows.net/test_dir/test_file.csv",
+            reader_method="csv",
+            reader_options={"header": True},
+            splitter_method="_split_on_whole_table",
+        )
+    ).dataframe
+    assert test_sparkdf.count() == 4
+    assert len(test_sparkdf.columns) == 2
+
+
+def test_get_batch_with_split_on_whole_table_gcs(
+    spark_session, basic_spark_df_execution_engine
+):
+    # noinspection PyUnusedLocal
+    def mocked_get_reader_function(*args, **kwargs):
+        # noinspection PyUnusedLocal,PyShadowingNames
+        def mocked_reader_function(*args, **kwargs):
+            pd_df = pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]})
+            df = spark_session.createDataFrame(
+                [
+                    tuple(
+                        None if isinstance(x, (float, int)) and np.isnan(x) else x
+                        for x in record.tolist()
+                    )
+                    for record in pd_df.to_records(index=False)
+                ],
+                pd_df.columns.tolist(),
+            )
+            return df
+
+        return mocked_reader_function
+
+    spark_engine = basic_spark_df_execution_engine
+    spark_engine._get_reader_fn = mocked_get_reader_function
+
+    test_sparkdf = spark_engine.get_batch_data(
+        GCSBatchSpec(
+            path="gcs://bucket/test/test.csv",
+            reader_method="csv",
+            reader_options={"header": True},
+            splitter_method="_split_on_whole_table",
+        )
+    ).dataframe
+    assert test_sparkdf.count() == 4
+    assert len(test_sparkdf.columns) == 2
+
+
+def test_get_batch_with_split_on_column_value(
+    test_sparkdf, basic_spark_df_execution_engine
+):
+    split_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf,
+            splitter_method="_split_on_column_value",
+            splitter_kwargs={
+                "column_name": "batch_id",
+                "batch_identifiers": {"batch_id": 2},
+            },
+        )
+    ).dataframe
+    assert test_sparkdf.count() == 120
+    assert len(test_sparkdf.columns) == 10
+    collected = split_df.collect()
+    for val in collected:
+        assert val.batch_id == 2
+
+    split_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf,
+            splitter_method="_split_on_column_value",
+            splitter_kwargs={
+                "column_name": "date",
+                "batch_identifiers": {"date": datetime.date(2020, 1, 30)},
+            },
+        )
+    ).dataframe
+    assert split_df.count() == 3
+    assert len(split_df.columns) == 10
+
+
+def test_get_batch_with_split_on_converted_datetime(
+    test_sparkdf, basic_spark_df_execution_engine
+):
+    split_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf,
+            splitter_method="_split_on_converted_datetime",
+            splitter_kwargs={
+                "column_name": "timestamp",
+                "batch_identifiers": {"timestamp": "2020-01-03"},
+            },
+        )
+    ).dataframe
+    assert split_df.count() == 2
+    assert len(split_df.columns) == 10
+
+
+def test_get_batch_with_split_on_divided_integer(
+    test_sparkdf, basic_spark_df_execution_engine
+):
+    split_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf,
+            splitter_method="_split_on_divided_integer",
+            splitter_kwargs={
+                "column_name": "id",
+                "divisor": 10,
+                "batch_identifiers": {"id": 5},
+            },
+        )
+    ).dataframe
+    assert split_df.count() == 10
+    assert len(split_df.columns) == 10
+    max_result = split_df.select([F.max("id")])
+    assert max_result.collect()[0]["max(id)"] == 59
+    min_result = split_df.select([F.min("id")])
+    assert min_result.collect()[0]["min(id)"] == 50
+
+
+def test_get_batch_with_split_on_mod_integer(
+    test_sparkdf, basic_spark_df_execution_engine
+):
+    split_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf,
+            splitter_method="_split_on_mod_integer",
+            splitter_kwargs={
+                "column_name": "id",
+                "mod": 10,
+                "batch_identifiers": {"id": 5},
+            },
+        )
+    ).dataframe
+
+    assert split_df.count() == 12
+    assert len(split_df.columns) == 10
+    max_result = split_df.select([F.max("id")])
+    assert max_result.collect()[0]["max(id)"] == 115
+    min_result = split_df.select([F.min("id")])
+    assert min_result.collect()[0]["min(id)"] == 5
+
+
+def test_get_batch_with_split_on_multi_column_values(
+    test_sparkdf, basic_spark_df_execution_engine
+):
+    split_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf,
+            splitter_method="_split_on_multi_column_values",
+            splitter_kwargs={
+                "column_names": ["y", "m", "d"],
+                "batch_identifiers": {
+                    "y": 2020,
+                    "m": 1,
+                    "d": 5,
+                },
+            },
+        )
+    ).dataframe
+    assert split_df.count() == 4
+    assert len(split_df.columns) == 10
+    collected = split_df.collect()
+    for val in collected:
+        assert val.date == datetime.date(2020, 1, 5)
+
+    with pytest.raises(ValueError):
+        # noinspection PyUnusedLocal
+        split_df = basic_spark_df_execution_engine.get_batch_data(
+            RuntimeDataBatchSpec(
+                batch_data=test_sparkdf,
+                splitter_method="_split_on_multi_column_values",
+                splitter_kwargs={
+                    "column_names": ["I", "dont", "exist"],
+                    "batch_identifiers": {
+                        "y": 2020,
+                        "m": 1,
+                        "d": 5,
+                    },
+                },
+            )
+        ).dataframe
+
+
+def test_get_batch_with_split_on_hashed_column_incorrect_hash_function_name(
+    test_sparkdf,
+    basic_spark_df_execution_engine,
+):
+    with pytest.raises(ge_exceptions.ExecutionEngineError):
+        # noinspection PyUnusedLocal
+        split_df = basic_spark_df_execution_engine.get_batch_data(
+            RuntimeDataBatchSpec(
+                batch_data=test_sparkdf,
+                splitter_method="_split_on_hashed_column",
+                splitter_kwargs={
+                    "column_name": "favorite_color",
+                    "hash_digits": 1,
+                    "hash_function_name": "I_wont_work",
+                    "batch_identifiers": {
+                        "hash_value": "a",
+                    },
+                },
+            )
+        ).dataframe
+
+
+def test_get_batch_with_split_on_hashed_column(
+    test_sparkdf, basic_spark_df_execution_engine
+):
+    split_df = basic_spark_df_execution_engine.get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_sparkdf,
+            splitter_method="_split_on_hashed_column",
+            splitter_kwargs={
+                "column_name": "favorite_color",
+                "hash_digits": 1,
+                "hash_function_name": "sha256",
+                "batch_identifiers": {
+                    "hash_value": "a",
+                },
+            },
+        )
+    ).dataframe
+    assert split_df.count() == 8
+    assert len(split_df.columns) == 10
