+++ tests/execution_engine/test_sparkdf_execution_engine.py
@@ -8,13 +8,7 @@ import pandas as pd
 import pytest
 
 import great_expectations.exceptions as ge_exceptions
-from great_expectations.core.batch_spec import (
-    AzureBatchSpec,
-    GCSBatchSpec,
-    PathBatchSpec,
-    RuntimeDataBatchSpec,
-    S3BatchSpec,
-)
+from great_expectations.core.batch_spec import PathBatchSpec, RuntimeDataBatchSpec
 from great_expectations.execution_engine import SparkDFExecutionEngine
 from great_expectations.execution_engine.execution_engine import MetricDomainTypes
 from great_expectations.expectations.row_conditions import (
@@ -38,70 +32,6 @@ except ImportError:
     LongType = None
     StringType = None
     Row = None
-
-
-@pytest.fixture
-def test_sparkdf(spark_session):
-    def generate_ascending_list_of_datetimes(
-        n, start_date=datetime.date(2020, 1, 1), end_date=datetime.date(2020, 12, 31)
-    ):
-        start_time = datetime.datetime(
-            start_date.year, start_date.month, start_date.day
-        )
-        seconds_between_dates = (end_date - start_date).total_seconds()
-        # noinspection PyUnusedLocal
-        datetime_list = [
-            start_time
-            + datetime.timedelta(seconds=random.randrange(int(seconds_between_dates)))
-            for i in range(n)
-        ]
-        datetime_list.sort()
-        return datetime_list
-
-    k = 120
-    random.seed(1)
-    timestamp_list = generate_ascending_list_of_datetimes(
-        n=k, end_date=datetime.date(2020, 1, 31)
-    )
-    date_list = [datetime.date(ts.year, ts.month, ts.day) for ts in timestamp_list]
-
-    # noinspection PyUnusedLocal
-    batch_ids = [random.randint(0, 10) for i in range(k)]
-    batch_ids.sort()
-    # noinspection PyUnusedLocal
-    session_ids = [random.randint(2, 60) for i in range(k)]
-    session_ids = [i - random.randint(0, 2) for i in session_ids]
-    session_ids.sort()
-
-    # noinspection PyUnusedLocal
-    spark_df = spark_session.createDataFrame(
-        data=pd.DataFrame(
-            {
-                "id": range(k),
-                "batch_id": batch_ids,
-                "date": date_list,
-                "y": [d.year for d in date_list],
-                "m": [d.month for d in date_list],
-                "d": [d.day for d in date_list],
-                "timestamp": timestamp_list,
-                "session_ids": session_ids,
-                "event_type": [
-                    random.choice(["start", "stop", "continue"]) for i in range(k)
-                ],
-                "favorite_color": [
-                    "#"
-                    + "".join(
-                        [random.choice(list("0123456789ABCDEF")) for j in range(6)]
-                    )
-                    for i in range(k)
-                ],
-            }
-        )
-    )
-    spark_df = spark_df.withColumn(
-        "timestamp", F.col("timestamp").cast(IntegerType()).cast(StringType())
-    )
-    return spark_df
 
 
 @pytest.fixture
@@ -670,322 +600,6 @@ def test_get_batch_empty_splitter_parquet(
     ).dataframe
     assert test_sparkdf.count() == 5
     assert len(test_sparkdf.columns) == 2
-
-
-def test_get_batch_with_split_on_whole_table_runtime(
-    test_sparkdf, basic_spark_df_execution_engine
-):
-    test_sparkdf = basic_spark_df_execution_engine.get_batch_data(
-        RuntimeDataBatchSpec(
-            batch_data=test_sparkdf, splitter_method="_split_on_whole_table"
-        )
-    ).dataframe
-    assert test_sparkdf.count() == 120
-    assert len(test_sparkdf.columns) == 10
-
-
-def test_get_batch_with_split_on_whole_table_filesystem(
-    test_folder_connection_path_csv, basic_spark_df_execution_engine
-):
-    # reader_method not configured because spark will configure own reader by default
-    test_sparkdf = basic_spark_df_execution_engine.get_batch_data(
-        PathBatchSpec(
-            path=os.path.join(test_folder_connection_path_csv, "test.csv"),
-            splitter_method="_split_on_whole_table",
-        )
-    ).dataframe
-    assert test_sparkdf.count() == 6
-    assert len(test_sparkdf.columns) == 2
-
-
-def test_get_batch_with_split_on_whole_table_s3(
-    spark_session, basic_spark_df_execution_engine
-):
-    # noinspection PyUnusedLocal
-    def mocked_get_reader_function(*args, **kwargs):
-        # noinspection PyUnusedLocal,PyShadowingNames
-        def mocked_reader_function(*args, **kwargs):
-            pd_df = pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]})
-            df = spark_session.createDataFrame(
-                [
-                    tuple(
-                        None if isinstance(x, (float, int)) and np.isnan(x) else x
-                        for x in record.tolist()
-                    )
-                    for record in pd_df.to_records(index=False)
-                ],
-                pd_df.columns.tolist(),
-            )
-            return df
-
-        return mocked_reader_function
-
-    spark_engine = basic_spark_df_execution_engine
-    spark_engine._get_reader_fn = mocked_get_reader_function
-
-    test_sparkdf = spark_engine.get_batch_data(
-        S3BatchSpec(
-            path="s3://bucket/test/test.csv",
-            reader_method="csv",
-            reader_options={"header": True},
-            splitter_method="_split_on_whole_table",
-        )
-    ).dataframe
-    assert test_sparkdf.count() == 4
-    assert len(test_sparkdf.columns) == 2
-
-
-def test_get_batch_with_split_on_whole_table_azure(
-    spark_session, basic_spark_df_execution_engine
-):
-    # noinspection PyUnusedLocal
-    def mocked_get_reader_function(*args, **kwargs):
-        # noinspection PyUnusedLocal,PyShadowingNames
-        def mocked_reader_function(*args, **kwargs):
-            pd_df = pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]})
-            df = spark_session.createDataFrame(
-                [
-                    tuple(
-                        None if isinstance(x, (float, int)) and np.isnan(x) else x
-                        for x in record.tolist()
-                    )
-                    for record in pd_df.to_records(index=False)
-                ],
-                pd_df.columns.tolist(),
-            )
-            return df
-
-        return mocked_reader_function
-
-    spark_engine = basic_spark_df_execution_engine
-    spark_engine._get_reader_fn = mocked_get_reader_function
-
-    test_sparkdf = spark_engine.get_batch_data(
-        AzureBatchSpec(
-            path="wasbs://test_container@test_account.blob.core.windows.net/test_dir/test_file.csv",
-            reader_method="csv",
-            reader_options={"header": True},
-            splitter_method="_split_on_whole_table",
-        )
-    ).dataframe
-    assert test_sparkdf.count() == 4
-    assert len(test_sparkdf.columns) == 2
-
-
-def test_get_batch_with_split_on_whole_table_gcs(
-    spark_session, basic_spark_df_execution_engine
-):
-    # noinspection PyUnusedLocal
-    def mocked_get_reader_function(*args, **kwargs):
-        # noinspection PyUnusedLocal,PyShadowingNames
-        def mocked_reader_function(*args, **kwargs):
-            pd_df = pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]})
-            df = spark_session.createDataFrame(
-                [
-                    tuple(
-                        None if isinstance(x, (float, int)) and np.isnan(x) else x
-                        for x in record.tolist()
-                    )
-                    for record in pd_df.to_records(index=False)
-                ],
-                pd_df.columns.tolist(),
-            )
-            return df
-
-        return mocked_reader_function
-
-    spark_engine = basic_spark_df_execution_engine
-    spark_engine._get_reader_fn = mocked_get_reader_function
-
-    test_sparkdf = spark_engine.get_batch_data(
-        GCSBatchSpec(
-            path="gcs://bucket/test/test.csv",
-            reader_method="csv",
-            reader_options={"header": True},
-            splitter_method="_split_on_whole_table",
-        )
-    ).dataframe
-    assert test_sparkdf.count() == 4
-    assert len(test_sparkdf.columns) == 2
-
-
-def test_get_batch_with_split_on_column_value(
-    test_sparkdf, basic_spark_df_execution_engine
-):
-    split_df = basic_spark_df_execution_engine.get_batch_data(
-        RuntimeDataBatchSpec(
-            batch_data=test_sparkdf,
-            splitter_method="_split_on_column_value",
-            splitter_kwargs={
-                "column_name": "batch_id",
-                "batch_identifiers": {"batch_id": 2},
-            },
-        )
-    ).dataframe
-    assert test_sparkdf.count() == 120
-    assert len(test_sparkdf.columns) == 10
-    collected = split_df.collect()
-    for val in collected:
-        assert val.batch_id == 2
-
-    split_df = basic_spark_df_execution_engine.get_batch_data(
-        RuntimeDataBatchSpec(
-            batch_data=test_sparkdf,
-            splitter_method="_split_on_column_value",
-            splitter_kwargs={
-                "column_name": "date",
-                "batch_identifiers": {"date": datetime.date(2020, 1, 30)},
-            },
-        )
-    ).dataframe
-    assert split_df.count() == 3
-    assert len(split_df.columns) == 10
-
-
-def test_get_batch_with_split_on_converted_datetime(
-    test_sparkdf, basic_spark_df_execution_engine
-):
-    split_df = basic_spark_df_execution_engine.get_batch_data(
-        RuntimeDataBatchSpec(
-            batch_data=test_sparkdf,
-            splitter_method="_split_on_converted_datetime",
-            splitter_kwargs={
-                "column_name": "timestamp",
-                "batch_identifiers": {"timestamp": "2020-01-03"},
-            },
-        )
-    ).dataframe
-    assert split_df.count() == 2
-    assert len(split_df.columns) == 10
-
-
-def test_get_batch_with_split_on_divided_integer(
-    test_sparkdf, basic_spark_df_execution_engine
-):
-    split_df = basic_spark_df_execution_engine.get_batch_data(
-        RuntimeDataBatchSpec(
-            batch_data=test_sparkdf,
-            splitter_method="_split_on_divided_integer",
-            splitter_kwargs={
-                "column_name": "id",
-                "divisor": 10,
-                "batch_identifiers": {"id": 5},
-            },
-        )
-    ).dataframe
-    assert split_df.count() == 10
-    assert len(split_df.columns) == 10
-    max_result = split_df.select([F.max("id")])
-    assert max_result.collect()[0]["max(id)"] == 59
-    min_result = split_df.select([F.min("id")])
-    assert min_result.collect()[0]["min(id)"] == 50
-
-
-def test_get_batch_with_split_on_mod_integer(
-    test_sparkdf, basic_spark_df_execution_engine
-):
-    split_df = basic_spark_df_execution_engine.get_batch_data(
-        RuntimeDataBatchSpec(
-            batch_data=test_sparkdf,
-            splitter_method="_split_on_mod_integer",
-            splitter_kwargs={
-                "column_name": "id",
-                "mod": 10,
-                "batch_identifiers": {"id": 5},
-            },
-        )
-    ).dataframe
-
-    assert split_df.count() == 12
-    assert len(split_df.columns) == 10
-    max_result = split_df.select([F.max("id")])
-    assert max_result.collect()[0]["max(id)"] == 115
-    min_result = split_df.select([F.min("id")])
-    assert min_result.collect()[0]["min(id)"] == 5
-
-
-def test_get_batch_with_split_on_multi_column_values(
-    test_sparkdf, basic_spark_df_execution_engine
-):
-    split_df = basic_spark_df_execution_engine.get_batch_data(
-        RuntimeDataBatchSpec(
-            batch_data=test_sparkdf,
-            splitter_method="_split_on_multi_column_values",
-            splitter_kwargs={
-                "column_names": ["y", "m", "d"],
-                "batch_identifiers": {
-                    "y": 2020,
-                    "m": 1,
-                    "d": 5,
-                },
-            },
-        )
-    ).dataframe
-    assert split_df.count() == 4
-    assert len(split_df.columns) == 10
-    collected = split_df.collect()
-    for val in collected:
-        assert val.date == datetime.date(2020, 1, 5)
-
-    with pytest.raises(ValueError):
-        # noinspection PyUnusedLocal
-        split_df = basic_spark_df_execution_engine.get_batch_data(
-            RuntimeDataBatchSpec(
-                batch_data=test_sparkdf,
-                splitter_method="_split_on_multi_column_values",
-                splitter_kwargs={
-                    "column_names": ["I", "dont", "exist"],
-                    "batch_identifiers": {
-                        "y": 2020,
-                        "m": 1,
-                        "d": 5,
-                    },
-                },
-            )
-        ).dataframe
-
-
-def test_get_batch_with_split_on_hashed_column_incorrect_hash_function_name(
-    test_sparkdf,
-    basic_spark_df_execution_engine,
-):
-    with pytest.raises(ge_exceptions.ExecutionEngineError):
-        # noinspection PyUnusedLocal
-        split_df = basic_spark_df_execution_engine.get_batch_data(
-            RuntimeDataBatchSpec(
-                batch_data=test_sparkdf,
-                splitter_method="_split_on_hashed_column",
-                splitter_kwargs={
-                    "column_name": "favorite_color",
-                    "hash_digits": 1,
-                    "hash_function_name": "I_wont_work",
-                    "batch_identifiers": {
-                        "hash_value": "a",
-                    },
-                },
-            )
-        ).dataframe
-
-
-def test_get_batch_with_split_on_hashed_column(
-    test_sparkdf, basic_spark_df_execution_engine
-):
-    split_df = basic_spark_df_execution_engine.get_batch_data(
-        RuntimeDataBatchSpec(
-            batch_data=test_sparkdf,
-            splitter_method="_split_on_hashed_column",
-            splitter_kwargs={
-                "column_name": "favorite_color",
-                "hash_digits": 1,
-                "hash_function_name": "sha256",
-                "batch_identifiers": {
-                    "hash_value": "a",
-                },
-            },
-        )
-    ).dataframe
-    assert split_df.count() == 8
-    assert len(split_df.columns) == 10
 
 
 # ### Sampling methods ###
